Create state_cache module
